Move st.error, st.info, st.success, st.warning test to new file
