+++ lib/tests/streamlit/streamlit_test.py
@@ -32,7 +32,6 @@ from parameterized import parameterized
 import streamlit as st
 from streamlit import __version__
 from streamlit.errors import StreamlitAPIException
-from streamlit.proto.Alert_pb2 import Alert
 from tests import testutil
 from tests.delta_generator_test_case import DeltaGeneratorTestCase
 
@@ -191,23 +190,6 @@ class StreamlitTest(unittest.TestCase):
 class StreamlitAPITest(DeltaGeneratorTestCase):
     """Test Public Streamlit Public APIs."""
 
-    def test_st_error(self):
-        """Test st.error."""
-        st.error("some error")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some error")
-        self.assertEqual(el.alert.format, Alert.ERROR)
-
-    def test_st_error_with_icon(self):
-        """Test st.error with icon."""
-        st.error("some error", icon="😱")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some error")
-        self.assertEqual(el.alert.icon, "😱")
-        self.assertEqual(el.alert.format, Alert.ERROR)
-
     @parameterized.expand([(True,), (False,)])
     def test_st_exception(self, show_error_details: bool):
         """Test st.exception."""
@@ -239,23 +221,6 @@ class StreamlitAPITest(DeltaGeneratorTestCase):
             el.doc_string.startswith("Change the current working directory")
         )
         self.assertEqual(f"posix.chdir(path)", el.value)
-
-    def test_st_info(self):
-        """Test st.info."""
-        st.info("some info")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some info")
-        self.assertEqual(el.alert.format, Alert.INFO)
-
-    def test_st_info_with_icon(self):
-        """Test st.info with icon."""
-        st.info("some info", icon="👉🏻")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some info")
-        self.assertEqual(el.alert.icon, "👉🏻")
-        self.assertEqual(el.alert.format, Alert.INFO)
 
     def test_st_json(self):
         """Test st.json."""
@@ -389,23 +354,6 @@ class StreamlitAPITest(DeltaGeneratorTestCase):
         self.assertNotEqual(el.plotly_chart.url, "the_url")
         self.assertEqual(el.plotly_chart.use_container_width, False)
 
-    def test_st_success(self):
-        """Test st.success."""
-        st.success("some success")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some success")
-        self.assertEqual(el.alert.format, Alert.SUCCESS)
-
-    def test_st_success_with_icon(self):
-        """Test st.success with icon."""
-        st.success("some success", icon="✅")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some success")
-        self.assertEqual(el.alert.icon, "✅")
-        self.assertEqual(el.alert.format, Alert.SUCCESS)
-
     def test_st_legacy_table(self):
         """Test st._legacy_table."""
         df = pd.DataFrame([[1, 2], [3, 4]], columns=["col1", "col2"])
@@ -501,23 +449,6 @@ class StreamlitAPITest(DeltaGeneratorTestCase):
         """Test st._legacy_vega_lite_chart."""
         pass
 
-    def test_st_warning(self):
-        """Test st.warning."""
-        st.warning("some warning")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some warning")
-        self.assertEqual(el.alert.format, Alert.WARNING)
-
-    def test_st_warning_with_icon(self):
-        """Test st.warning with icon."""
-        st.warning("some warning", icon="⚠️")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.alert.body, "some warning")
-        self.assertEqual(el.alert.icon, "⚠️")
-        self.assertEqual(el.alert.format, Alert.WARNING)
-
     def test_set_query_params_sends_protobuf_message(self):
         """Test valid st.set_query_params sends protobuf message."""
         st.experimental_set_query_params(x="a")
@@ -537,9 +468,3 @@ class StreamlitAPITest(DeltaGeneratorTestCase):
         st.experimental_set_query_params(**p_set)
         p_get = st.experimental_get_query_params()
         self.assertEqual(p_get, p_set)
-
-    @parameterized.expand([(st.error,), (st.warning,), (st.info,), (st.success,)])
-    def test_st_alert_exceptions(self, alert_func):
-        """Test that alert functions throw an exception when a non-emoji is given as an icon."""
-        with self.assertRaises(StreamlitAPIException):
-            alert_func("some alert", icon="hello world")
